Comment and User are moved to the Entity package
