+++ src/main/java/ru/skypro/homework/entity/Comment.java
@@ -1,4 +1,4 @@
-package ru.skypro.homework.model;
+package ru.skypro.homework.entity;
 
 import javax.persistence.*;
 import java.time.Instant;
