+++ src/main/java/ru/skypro/homework/mapper/UserMapper.java
@@ -2,13 +2,10 @@ package ru.skypro.homework.mapper;
 
 import org.mapstruct.IterableMapping;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
 import ru.skypro.homework.dto.User.UpdateUserDTO;
 import ru.skypro.homework.dto.User.UserDTO;
-import ru.skypro.homework.dto.comments.CreateOrUpdateCommentDTO;
-import ru.skypro.homework.model.Comment;
-import ru.skypro.homework.model.User;
+import ru.skypro.homework.entity.User;
 
 import java.util.List;
