+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -1,7 +1,7 @@
 package ru.skypro.homework.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.skypro.homework.model.Comment;
+import ru.skypro.homework.entity.Comment;
 
 public interface CommentRepository extends JpaRepository<Comment, Integer> {
 }
